feat: refactor MIAlert and MIChip components for improved accessibility and remove unused props; update theme and types
